% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{calc_fies}
\alias{calc_fies}
\title{Score FIES (Securite Alimentaire)}
\usage{
calc_fies(s08, poids = NULL)
}
\arguments{
\item{s08}{La base s08a_me (section securite alimentaire)}

\item{poids}{Base welfare contenant hhweight (optionnel)}
}
\value{
Un data.frame avec hhid, score_fies, fies_moder, fies_sever
}
\description{
Construit le score FIES (0-8) et les seuils d'insecurite moderee/severe
a partir des 8 questions dichotomiques de la section 8.
}
