% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{calc_hdds}
\alias{calc_hdds}
\title{Score HDDS (Diversite Alimentaire)}
\usage{
calc_hdds(s07b, passage_fao)
}
\arguments{
\item{s07b}{La base s07b_me (consommation alimentaire)}

\item{passage_fao}{Table de passage code produit -> groupe FAO (12 groupes)}
}
\value{
Un data.frame avec hhid et hdds (score 0-12)
}
\description{
Calcule le Household Dietary Diversity Score (nombre de groupes alimentaires
FAO consommes sur 7 jours) a partir de la section 7B.
}
