% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{calc_rendement}
\alias{calc_rendement}
\title{Calcul du Rendement (kg/ha)}
\usage{
calc_rendement(data, codes_prod, table_conv)
}
\arguments{
\item{data}{L'objet renvoye par load_filiere()}

\item{codes_prod}{Vecteur des codes culture dans s16c}

\item{table_conv}{La table de conversion (phase 2) pour les unites locales}
}
\value{
Un data.frame avec hhid, surface_ha, production_kg, rendement_kg_ha, hhweight
}
\description{
Calcule le rendement (kg/ha) avec conversion des unites locales et winsorisation.
}
