% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{carte_filiere}
\alias{carte_filiere}
\title{Carte leaflet d'un indicateur de la filiere}
\usage{
carte_filiere(data, indicateur, nom_col, titre)
}
\arguments{
\item{data}{L'objet renvoye par load_filiere()}

\item{indicateur}{Un data.frame avec grappe + une colonne numerique a cartographier}

\item{nom_col}{Le nom de la colonne numerique a afficher}

\item{titre}{Le titre de la carte}
}
\value{
Un widget leaflet
}
\description{
Genere une carte interactive (leaflet) d'un indicateur au niveau grappe.
}
