% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\docType{package}
\name{filiereBFA-package}
\alias{filiereBFA}
\alias{filiereBFA-package}
\title{filiereBFA : Analyse de filiere et securite alimentaire (EHCVM)}
\description{
Ce package encapsule la chaine analytique du projet filiere EHCVM.
Il permet de realiser les Modules 1 a 5 du projet ISEP2 pour n'importe
quel produit strategique (mais, mil, sorgho...) et pays UEMOA.
}
\section{Fonctions principales }{


\itemize{
  \item \code{\link{load_filiere}} : charge et fusionne les bases EHCVM
  \item \code{\link{calc_rendement}} : calcul du rendement (kg/ha)
  \item \code{\link{calc_fies}} : construction du score FIES
  \item \code{\link{calc_hdds}} : score de diversite alimentaire
  \item \code{\link{profil_menage}} : profilage socioeconomique des 4 groupes
  \item \code{\link{prix_chaine}} : prix producteur vs consommation + marge
  \item \code{\link{carte_filiere}} : carte leaflet d'un indicateur
  \item \code{\link{reg_filiere}} : regression de l'impact filiere sur FIES/HDDS
}
}

