% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{load_filiere}
\alias{load_filiere}
\title{Importation et fusion des bases EHCVM}
\usage{
load_filiere(dossier, produit, codes_prod, codes_conso, pays = "BFA")
}
\arguments{
\item{dossier}{Le chemin vers le dossier contenant les fichiers .dta}

\item{produit}{Le nom du produit strategique (ex: "maïs", "mil")}

\item{codes_prod}{Vecteur des codes culture dans s16c (ex: 4L pour le mais)}

\item{codes_conso}{Vecteur des codes produit conso dans s07b (ex: c(5,6,12,13))}

\item{pays}{Code pays (ex: "BFA" pour le Burkina Faso)}
}
\value{
Un objet de type liste contenant les bases chargees et la typologie des menages.
}
\description{
Charge les sections EHCVM (menage + communautaire) pour un pays et un produit
donnes. Cree un objet "data.filiere" contenant les bases pretes pour l'analyse.
}
