% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{prix_chaine}
\alias{prix_chaine}
\title{Chaine de Prix (Commercialisation)}
\usage{
prix_chaine(data, codes_prod, codes_conso, table_conv_s07)
}
\arguments{
\item{data}{L'objet renvoye par load_filiere()}

\item{codes_prod}{Vecteur des codes culture dans s16c/s16d}

\item{codes_conso}{Vecteur des codes produit conso dans s07b}

\item{table_conv_s07}{La table de conversion pour s07b (methode Cours 7)}
}
\value{
Une liste contenant prix_prod, prix_conso, et marge_region
}
\description{
Calcule les prix producteur et consommation, ainsi que la marge commerciale.
}
