% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filiereBFA-package.r
\name{profil_menage}
\alias{profil_menage}
\title{Profil socioeconomique des 4 groupes de menages}
\usage{
profil_menage(data, fies, hdds)
}
\arguments{
\item{data}{L'objet renvoye par load_filiere()}

\item{fies}{La base FIES calculee par calc_fies()}

\item{hdds}{La base HDDS calculee par calc_hdds()}
}
\value{
Un data.frame groupant les indicateurs par groupe de menage.
}
\description{
Genere un tableau comparatif des 4 groupes (Producteur-Conso, Producteur seul,
Conso seul, Ni prod ni conso) sur des indicateurs cles (age, sexe, education
du chef, pauvrete, FIES, HDDS).
}
